#2183 merged r10859 into stable
git-svn-id: https://src.heuristiclab.com/svn/core/stable@11164 2abd9481-f8db-48e9-bd25-06bc13291c1b

+++ HeuristicLab.Encodings.SymbolicExpressionTreeEncoding/3.4/Compiler/LinearInstruction.cs
@@ -24,7 +24,7 @@ namespace HeuristicLab.Encodings.SymbolicExpressionTreeEncoding {
   // total size of this class should be small to improve cache access while executing the code
   public class LinearInstruction : Instruction {
     public double value;
-    public byte childIndex;
+    public int childIndex;
     public bool skip;
   }
 }
